Refactor DeviceConfig to encapsulate device capabilities, enhancing clarity and functionality in managing audio device configurations.
